+++ mrbgems/mruby-eval/mrbgem.rake
@@ -5,4 +5,6 @@ MRuby::Gem::Specification.new('mruby-eval') do |spec|
 
   add_dependency 'mruby-compiler', :core => 'mruby-compiler'
   add_dependency 'mruby-binding', :core => 'mruby-binding'
+  spec.add_test_dependency('mruby-metaprog', :core => 'mruby-metaprog')
+  spec.add_test_dependency('mruby-method', :core => 'mruby-method')
 end
