+++ mrbgems/mruby-eval/test/binding.rb
