+++ mrbgems/mruby-proc-binding/mrbgem.rake
@@ -5,6 +5,6 @@ MRuby::Gem::Specification.new('mruby-proc-binding') do |spec|
 
   spec.add_dependency('mruby-binding', :core => 'mruby-binding')
   spec.add_dependency('mruby-proc-ext', :core => 'mruby-proc-ext')
-  spec.add_test_dependency('mruby-binding-eval', :core => 'mruby-binding-eval')
+  spec.add_test_dependency('mruby-eval', :core => 'mruby-eval')
   spec.add_test_dependency('mruby-compiler', :core => 'mruby-compiler')
 end
